--- app/client/static/UserCourse.ts
@@ -1,11 +1,0 @@
-import Course from "./Course";
-import Exam from "./Exam";
-
-export default interface UserCourse{
-    course: Course,
-    isLK: boolean,
-    isWrittenExamCourse: boolean,
-    isOralExamCourse: boolean,
-    writtenWeightPercentage: number,
-    exams: Exam[],
-}
