+++ app/client/static/interfaces/Course.ts
