+++ app/client/static/interfaces/Exam.ts
