+++ app/client/static/interfaces/PointWeights.ts
